# internal/httpapi/openapi.yaml
openapi: 3.0.3
info:
  title: Garfex API
  version: 0.1.0
  description: Public operational, API-reference, catalog-descriptor, catalog-list, and catalog-detail endpoints.
paths:
...
  /v1/resources:
    get:
      summary: Search resources
      description: Returns a page of resources matching the query.
      parameters:
        - name: scope
          in: query
          schema: { type: string, enum: [ACTIVE, INACTIVE, ALL], default: ACTIVE }
        - name: text
          in: query
          schema: { type: string }
        - name: classCode
          in: query
          schema: { type: string }
        - name: familyCode
          in: query
          schema: { type: string }
        - name: typeCode
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        '200':
          description: Resource page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourcePage'
        '400': { $ref: '#/components/responses/InvalidArgument' }
        '500': { $ref: '#/components/responses/Internal' }
    post:
      summary: Create a resource
      description: Creates one resource under a class/family/type scope.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceCreateRequest'
      responses:
        '201':
          description: Created resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
        '400': { $ref: '#/components/responses/InvalidArgument' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '503': { $ref: '#/components/responses/Unavailable' }
        '500': { $ref: '#/components/responses/Internal' }
  /v1/resources/{classCode}/{identityV1}:
    get:
      summary: Get a resource
      description: Returns one resource by class code and durable identity-v1.
      parameters:
        - name: classCode
          in: path
          required: true
          schema: { type: string }
        - name: identityV1
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
        '400': { $ref: '#/components/responses/InvalidArgument' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/Internal' }
...
  /v1/resources/{id}:
    put:
      summary: Update a resource
      description: Replaces one existing resource's content by its internal id, obtained from a prior read, under optimistic concurrency.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, description: 'Strict positive decimal identifier; the runtime rejects values exceeding the int64 boundary 9223372036854775807.', pattern: '^[1-9][0-9]*$', maxLength: 19, not: { pattern: '[\r\n]' } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceUpdateRequest'
      responses:
        '200':
          description: Updated resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
        '400': { $ref: '#/components/responses/InvalidArgument' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '500': { $ref: '#/components/responses/Internal' }
...
components:
  responses:
    InvalidArgument: { description: Invalid request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    NotFound: { description: Requested resource was not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Conflict: { description: Request conflicts with current state, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Unprocessable: { description: Request failed validation, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Unavailable: { description: Service is temporarily unavailable, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Internal: { description: Internal server error, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  schemas:
...
    CatalogValue:
      oneOf:
        - { $ref: '#/components/schemas/CatalogTextValue' }
        - { $ref: '#/components/schemas/CatalogCodeValue' }
        - { $ref: '#/components/schemas/CatalogBooleanValue' }
        - { $ref: '#/components/schemas/CatalogIntegerValue' }
        - { $ref: '#/components/schemas/CatalogDecimalValue' }
        - { $ref: '#/components/schemas/CatalogQuantityValue' }
        - { $ref: '#/components/schemas/CatalogReferenceValue' }
        - { $ref: '#/components/schemas/CatalogEnumValue' }
        - { $ref: '#/components/schemas/CatalogStringListValue' }
        - { $ref: '#/components/schemas/CatalogControlledOptionValue' }
        - { $ref: '#/components/schemas/CatalogNotApplicableValue' }
      discriminator:
        propertyName: kind
        mapping:
          TEXT: '#/components/schemas/CatalogTextValue'
          CODE: '#/components/schemas/CatalogCodeValue'
          BOOLEAN: '#/components/schemas/CatalogBooleanValue'
          INTEGER: '#/components/schemas/CatalogIntegerValue'
          DECIMAL: '#/components/schemas/CatalogDecimalValue'
          QUANTITY: '#/components/schemas/CatalogQuantityValue'
          REFERENCE: '#/components/schemas/CatalogReferenceValue'
          ENUM: '#/components/schemas/CatalogEnumValue'
          STRING_LIST: '#/components/schemas/CatalogStringListValue'
          CONTROLLED_OPTION: '#/components/schemas/CatalogControlledOptionValue'
          NOT_APPLICABLE: '#/components/schemas/CatalogNotApplicableValue'
    CatalogTextValue:
      type: object
      additionalProperties: false
      required: [kind, value]
      properties: { kind: { type: string, enum: [TEXT] }, value: { type: string } }
    CatalogCodeValue:
      type: object
      additionalProperties: false
      required: [kind, value]
      properties: { kind: { type: string, enum: [CODE] }, value: { type: string } }
    CatalogBooleanValue:
      type: object
      additionalProperties: false
      required: [kind, value]
      properties: { kind: { type: string, enum: [BOOLEAN] }, value: { type: boolean } }
    CatalogIntegerValue:
      type: object
      additionalProperties: false
      required: [kind, value]
      properties: { kind: { type: string, enum: [INTEGER] }, value: { type: string, description: Decimal integer, pattern: '^(?:0|-?[1-9][0-9]*)$', not: { pattern: '[\r\n]' } } }
    CatalogDecimalValue:
      type: object
      additionalProperties: false
      required: [kind, value]
      properties: { kind: { type: string, enum: [DECIMAL] }, value: { type: string, pattern: '^(?:0|-?(?:[1-9][0-9]*|(?:0|[1-9][0-9]*)\.[0-9]*[1-9]))$', not: { pattern: '[\r\n]' } } }
    CatalogQuantityValue:
      type: object
      additionalProperties: false
      required: [kind, value, unitCode]
      properties: { kind: { type: string, enum: [QUANTITY] }, value: { type: string, pattern: '^(?:0|-?(?:[1-9][0-9]*|(?:0|[1-9][0-9]*)\.[0-9]*[1-9]))$', not: { pattern: '[\r\n]' } }, unitCode: { type: string } }
    CatalogReferenceValue:
      type: object
      additionalProperties: false
      required: [kind, reference]
      properties: { kind: { type: string, enum: [REFERENCE] }, reference: { $ref: '#/components/schemas/CatalogReference' } }
    CatalogReference:
      type: object
      additionalProperties: false
      required: [kind, id, code]
      properties:
        kind: { type: string }
        id:
          type: string
          pattern: '^(?:0|-?[1-9][0-9]*)$'
          not: { pattern: '[\r\n]' }
          description: >-
            Decimal integer identifier. In a response this is the referenced
            record's real id. In a create/update request body it MUST be "0":
            the server resolves the reference by code, and a non-zero id is
            rejected as invalid (references are addressed by natural code on
            write, never by id).
        code: { type: string }
    CatalogEnumValue:
      type: object
      additionalProperties: false
      required: [kind, value]
      properties: { kind: { type: string, enum: [ENUM] }, value: { type: string } }
    CatalogStringListValue:
      type: object
      additionalProperties: false
      required: [kind, values]
      properties: { kind: { type: string, enum: [STRING_LIST] }, values: { type: array, items: { type: string } } }
    CatalogControlledOptionValue:
      type: object
      additionalProperties: false
      required: [kind, value]
      properties: { kind: { type: string, enum: [CONTROLLED_OPTION] }, value: { type: string } }
    CatalogNotApplicableValue:
      type: object
      additionalProperties: false
      required: [kind]
      properties: { kind: { type: string, enum: [NOT_APPLICABLE] } }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: 'Generic, sanitized classification of the failure (e.g. "invalid request", "validation failed").'
        code:
          type: string
          description: >-
            Stable, machine-readable reason code, safe to switch on: distinct
            outcomes that read the same generic "conflict" error (e.g. an
            expected-revision mismatch, a still-active record, or a record
            still in use) are always distinguishable here even when error is
            identical. Absent only when the failure never reached Core (e.g.
            an HTTP method mismatch).
          enum:
            - INVALID_ARGUMENT
            - NOT_FOUND
            - DUPLICATE
            - INVALID_REFERENCE
            - VALIDATION
            - INTEGRITY
            - IDENTITY_CONFLICT
            - INVALID_LIFECYCLE
            - REACTIVATION_IMPOSSIBLE
            - INVALID_CATALOG
            - IN_USE
            - IMMUTABLE_CODE
            - CONFLICT
            - UNAVAILABLE
            - INTERNAL
        detail:
          type: string
          description: >-
            The specific reason, present only when error is "invalid request" or
            "validation failed". Absent for every other error (not found, conflict,
            unavailable, internal server error), which never expose internal detail.
...
    ResourceScope:
      type: object
      additionalProperties: false
      required: [classCode, familyCode, typeCode]
      properties:
        classCode: { type: string }
        familyCode: { type: string }
        typeCode: { type: string }
    ResourceAttribute:
      type: object
      additionalProperties: false
      required: [code, value]
      properties:
        code: { type: string }
        value: { $ref: '#/components/schemas/CatalogValue' }
    ResourceCreateRequest:
      type: object
      additionalProperties: false
      required: [actor, scope, naturalUnit, attributes]
      properties:
        actor: { type: string }
        scope: { $ref: '#/components/schemas/ResourceScope' }
        naturalUnit: { type: string }
        attributes: { type: array, items: { $ref: '#/components/schemas/ResourceAttribute' } }
    ResourceUpdateRequest:
      type: object
      additionalProperties: false
      required: [actor, expectedRevision, scope, naturalUnit, attributes]
      properties:
        actor: { type: string }
        expectedRevision: { type: string, description: Decimal unsigned revision, pattern: '^(?:0|[1-9][0-9]*)$', not: { pattern: '[\r\n]' } }
        scope: { $ref: '#/components/schemas/ResourceScope' }
        naturalUnit: { type: string }
        attributes: { type: array, items: { $ref: '#/components/schemas/ResourceAttribute' } }
...
    Resource:
      type: object
      additionalProperties: false
      required: [id, identityV1, scope, naturalUnit, active, revision, attributes]
      properties:
        id: { type: string, description: Decimal integer identifier, pattern: '^(?:0|-?[1-9][0-9]*)$', not: { pattern: '[\r\n]' } }
        identityV1: { type: string }
        scope: { $ref: '#/components/schemas/ResourceScope' }
        naturalUnit: { type: string }
        active: { type: boolean }
        revision: { type: string, description: Decimal unsigned revision, pattern: '^(?:0|[1-9][0-9]*)$', not: { pattern: '[\r\n]' } }
        attributes: { type: array, items: { $ref: '#/components/schemas/ResourceAttribute' } }
    ResourcePage:
      type: object
      additionalProperties: false
      required: [resources, hasPrevious, hasNext]
      properties:
        resources: { type: array, items: { $ref: '#/components/schemas/Resource' } }
        hasPrevious: { type: boolean }
        hasNext: { type: boolean }
...